+++ app/opulence/views/View.php
@@ -28,10 +28,6 @@ class View implements IView
     protected $path = "";
     /** @var array The mapping of PHP variable names to their values */
     protected $vars = [];
-    /** @var array The mapping of view part names to their contents */
-    protected $parts = [];
-    /** @var IView|null The parent view if there is one, otherwise false */
-    protected $parent;
     /** @var array The stack of parent delimiter types to values */
     private $delimiters = [
         self::DELIMITER_TYPE_UNSANITIZED_TAG => [
@@ -82,55 +78,6 @@ class View implements IView
     /**
      * @inheritdoc
      */
-    public function getParent()
-    {
-        return $this->parent;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function getPart($name)
-    {
-        if(isset($this->parts[$name]))
-        {
-            return $this->parts[$name];
-        }
-        elseif($this->parent !== null)
-        {
-            return $this->parent->getPart($name);
-        }
-
-        return "";
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function getParts()
-    {
-        $parts = $this->parts;
-        $currParent = $this->parent;
-
-        while($currParent !== null)
-        {
-            foreach($this->parent->getParts() as $name => $content)
-            {
-                if(!array_key_exists($name, $this->parts))
-                {
-                    $parts[$name] = $content;
-                }
-            }
-
-            $currParent = $this->parent->getParent();
-        }
-
-        return $parts;
-    }
-
-    /**
-     * @inheritdoc
-     */
     public function getPath()
     {
         return $this->path;
@@ -145,10 +92,6 @@ class View implements IView
         {
             return $this->vars[$name];
         }
-        elseif($this->parent !== null)
-        {
-            return $this->parent->getVar($name);
-        }
 
         return null;
     }
@@ -158,23 +101,15 @@ class View implements IView
      */
     public function getVars()
     {
-        $vars = $this->vars;
-        $currParent = $this->parent;
+        return $this->vars;
+    }
 
-        while($currParent !== null)
-        {
-            foreach($this->parent->getVars() as $name => $value)
-            {
-                if(!array_key_exists($name, $this->vars))
-                {
-                    $vars[$name] = $value;
-                }
-            }
-
-            $currParent = $this->parent->getParent();
-        }
-
-        return $vars;
+    /**
+     * @inheritdoc
+     */
+    public function hasVar($name)
+    {
+        return array_key_exists($name, $this->vars);
     }
 
     /**
@@ -200,33 +135,6 @@ class View implements IView
 
     /**
      * @inheritdoc
-     */
-    public function setParent(IView $parent)
-    {
-        $this->parent = $parent;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function setPart($name, $content)
-    {
-        $this->parts[$name] = $content;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function setParts(array $namesToContents)
-    {
-        foreach($namesToContents as $name => $content)
-        {
-            $this->setPart($name, $content);
-        }
-    }
-
-    /**
-     * @inheritDoc
      */
     public function setPath($path)
     {