+++ CHANGELOG.md
@@ -13,8 +13,7 @@
 * Renamed `RDevPredis` to `OpulencePredis`
 
 <h3>Views</h3>
-* Added `Opulence\Views\View` base class
-* Added `Opulence\Views\FortuneView` for views that use the Fortune engine
+* Added `Opulence\Views\View`
 * Renamed `Opulence\Views\Factories\ITemplateFactory` to `IViewFactory` and `TemplateFactory` to `ViewFactory`
 * Changed default statement delimiter from `{% %}` to `<% %>`